Revert

+++ Source/childcare-server/src/main/java/com/polimi/childcare/server/handlers/entities/setters/GitaRequestHandlerSet.java
@@ -1,7 +1,6 @@
 package com.polimi.childcare.server.handlers.entities.setters;
 
 import com.polimi.childcare.server.database.DatabaseSession;
-import com.polimi.childcare.server.helper.DBHelper;
 import com.polimi.childcare.shared.entities.Gita;
 import com.polimi.childcare.shared.entities.PianoViaggi;
 import com.polimi.childcare.shared.networking.requests.setters.SetEntityRequest;
@@ -26,16 +25,16 @@ public class GitaRequestHandlerSet extends GenericSetEntityRequestHandler<SetEnt
                 throw new RuntimeException("Un campo obbligatorio è null!");
 
             //Se ho rimosso dei piano viaggi, allora li cancello
-            /*if(dbEntity != null)
+            if(dbEntity != null)
             {
                 for(PianoViaggi pianoViaggi : dbEntity.getPianiViaggi())
                 {
                     if(!request.getEntity().getPianiViaggi().contains(pianoViaggi))
                         session.delete(pianoViaggi);
                 }
-            }*/
+            }
         }
-        /*else
+        else
         {
             //Rimuovo i piano viaggi legati a questo gruppo
             for(PianoViaggi viaggi : request.getEntity().getPianiViaggi())
@@ -48,6 +47,6 @@ public class GitaRequestHandlerSet extends GenericSetEntityRequestHandler<SetEnt
                 piano.setGruppo(null);
                 session.delete(piano);
             }
-        }*/
+        }
     }
 }
